+++ okio-testing-support/src/commonMain/kotlin/okio/AbstractFileSystemTest.kt
@@ -28,12 +28,9 @@ import kotlin.test.fail
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.minutes
 import kotlin.time.Duration.Companion.seconds
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
 import okio.ByteString.Companion.encodeUtf8
 import okio.ByteString.Companion.toByteString
 import okio.Path.Companion.toPath
-import okio.fakefilesystem.FakeFileSystem
 
 /** This test assumes that okio-files/ is the current working directory when executed. */
 abstract class AbstractFileSystemTest(
@@ -47,6 +44,7 @@ abstract class AbstractFileSystemTest(
 ) {
   val base: Path = temporaryDirectory / "${this::class.simpleName}-${randomToken(16)}"
   private val isNodeJsFileSystem = fileSystem::class.simpleName?.startsWith("NodeJs") ?: false
+  private val isWasiFileSystem = fileSystem::class.simpleName?.startsWith("Wasi") ?: false
   private val isWrappingJimFileSystem = this::class.simpleName?.contains("JimFileSystem") ?: false
 
   @BeforeTest
@@ -68,13 +66,15 @@ abstract class AbstractFileSystemTest(
 
   @Test
   fun canonicalizeDotReturnsCurrentWorkingDirectory() {
-    if (fileSystem is FakeFileSystem || fileSystem is ForwardingFileSystem) return
+    if (fileSystem.isFakeFileSystem || fileSystem is ForwardingFileSystem) return
     val cwd = fileSystem.canonicalize(".".toPath())
     val cwdString = cwd.toString()
     val slash = Path.DIRECTORY_SEPARATOR
     assertTrue(cwdString) {
       if (isWrappingJimFileSystem) {
         cwdString.endsWith("work")
+      } else if (isWasiFileSystem) {
+        cwdString.endsWith("/tmp")
       } else {
         cwdString.endsWith("okio${slash}okio") ||
           cwdString.endsWith("${slash}okio-parent-okio-nodefilesystem-test") ||
@@ -262,14 +262,14 @@ abstract class AbstractFileSystemTest(
   @Test
   fun listOnRelativePathReturnsRelativePaths() {
     // Make sure there's always at least one file so our assertion is useful.
-    if (fileSystem is FakeFileSystem) {
+    if (fileSystem.isFakeFileSystem) {
       val workingDirectory = "/directory".toPath()
       fileSystem.createDirectory(workingDirectory)
       fileSystem.workingDirectory = workingDirectory
       fileSystem.write("a.txt".toPath()) {
         writeUtf8("hello, world!")
       }
-    } else if (isWrappingJimFileSystem) {
+    } else if (isWrappingJimFileSystem || isWasiFileSystem) {
       fileSystem.write("a.txt".toPath()) {
         writeUtf8("hello, world!")
       }
@@ -286,20 +286,18 @@ abstract class AbstractFileSystemTest(
     // Make sure there's always at least one file so our assertion is useful. We copy the first 2
     // entries of the real working directory of the JVM to validate the results on all environment.
     if (
-      fileSystem is FakeFileSystem ||
-      fileSystem is ForwardingFileSystem && fileSystem.delegate is FakeFileSystem
+      fileSystem.isFakeFileSystem ||
+      fileSystem is ForwardingFileSystem && fileSystem.delegate.isFakeFileSystem
     ) {
       val workingDirectory = "/directory".toPath()
       fileSystem.createDirectory(workingDirectory)
-      (fileSystem as? FakeFileSystem)?.workingDirectory = workingDirectory
-      ((fileSystem as? ForwardingFileSystem)?.delegate as? FakeFileSystem)?.workingDirectory =
-        workingDirectory
+      fileSystem.workingDirectory = workingDirectory
       val apiDir = "api".toPath()
       fileSystem.createDirectory(apiDir)
       fileSystem.write(apiDir / "okio.api".toPath()) {
         writeUtf8("hello, world!")
       }
-    } else if (isWrappingJimFileSystem) {
+    } else if (isWrappingJimFileSystem || isWasiFileSystem) {
       val apiDir = "api".toPath()
       fileSystem.createDirectory(apiDir)
       fileSystem.write(apiDir / "okio.api".toPath()) {
@@ -339,14 +337,12 @@ abstract class AbstractFileSystemTest(
     // Make sure there's always at least one file so our assertion is useful. We copy the first 2
     // entries of the real working directory of the JVM to validate the results on all environment.
     if (
-      fileSystem is FakeFileSystem ||
-      fileSystem is ForwardingFileSystem && fileSystem.delegate is FakeFileSystem
+      fileSystem.isFakeFileSystem ||
+      fileSystem is ForwardingFileSystem && fileSystem.delegate.isFakeFileSystem
     ) {
       val workingDirectory = "/directory".toPath()
       fileSystem.createDirectory(workingDirectory)
-      (fileSystem as? FakeFileSystem)?.workingDirectory = workingDirectory
-      ((fileSystem as? ForwardingFileSystem)?.delegate as? FakeFileSystem)?.workingDirectory =
-        workingDirectory
+      fileSystem.workingDirectory = workingDirectory
       val apiDir = "api".toPath()
       fileSystem.createDirectory(apiDir)
       fileSystem.write(apiDir / "okio.api".toPath()) {
@@ -431,7 +427,7 @@ abstract class AbstractFileSystemTest(
   @Test
   fun listOrNullOnRelativePathReturnsRelativePaths() {
     // Make sure there's always at least one file so our assertion is useful.
-    if (fileSystem is FakeFileSystem) {
+    if (fileSystem.isFakeFileSystem) {
       val workingDirectory = "/directory".toPath()
       fileSystem.createDirectory(workingDirectory)
       fileSystem.workingDirectory = workingDirectory
@@ -918,7 +914,7 @@ abstract class AbstractFileSystemTest(
 
   @Test
   fun appendingSinkDoesNotImpactExistingFile() {
-    if (fileSystem is FakeFileSystem && !fileSystem.allowReadsWhileWriting) return
+    if (fileSystem.isFakeFileSystem && !fileSystem.allowReadsWhileWriting) return
 
     val path = base / "appending-sink-does-not-impact-existing-file"
     path.writeUtf8("hello, world!\n")
@@ -962,7 +958,7 @@ abstract class AbstractFileSystemTest(
 
   @Test
   fun fileSinkFlush() {
-    if (fileSystem is FakeFileSystem && !fileSystem.allowReadsWhileWriting) return
+    if (fileSystem.isFakeFileSystem && !fileSystem.allowReadsWhileWriting) return
 
     val path = base / "file-sink"
     val sink = fileSystem.sink(path)
@@ -2307,7 +2303,14 @@ abstract class AbstractFileSystemTest(
     val maxTime = clock.now()
 
     val sourceMetadata = fileSystem.metadata(source)
-    assertEquals(target, sourceMetadata.symlinkTarget)
+    // Okio's WasiFileSystem only creates relative symlinks.
+    assertEquals(
+      when {
+        isWasiFileSystem -> target.relativeTo(source.parent!!)
+        else -> target
+      },
+      sourceMetadata.symlinkTarget,
+    )
     assertInRange(sourceMetadata.createdAt, minTime, maxTime)
   }
 
@@ -2481,7 +2484,14 @@ abstract class AbstractFileSystemTest(
     fileSystem.atomicMove(source1, source2)
     assertEquals("I am the target file", target.readUtf8())
     assertEquals("I am the target file", source2.readUtf8())
-    assertEquals(target, fileSystem.metadata(source2).symlinkTarget)
+    // Okio's WasiFileSystem only creates relative symlinks.
+    assertEquals(
+      when {
+        isWasiFileSystem -> target.relativeTo(source1.parent!!)
+        else -> target
+      },
+      fileSystem.metadata(source2).symlinkTarget,
+    )
   }
 
   @Test
@@ -2544,7 +2554,7 @@ abstract class AbstractFileSystemTest(
   }
 
   protected fun supportsSymlink(): Boolean {
-    if (fileSystem is FakeFileSystem) return fileSystem.allowSymlinks
+    if (fileSystem.isFakeFileSystem) return fileSystem.allowSymlinks
     if (windowsLimitations) return false
     return when (fileSystem::class.simpleName) {
       "JvmSystemFileSystem",
@@ -2592,7 +2602,7 @@ abstract class AbstractFileSystemTest(
    */
   private fun Instant.minFileSystemTime(): Instant {
     val paddedInstant = minus(200.milliseconds)
-    return Instant.fromEpochSeconds(paddedInstant.epochSeconds)
+    return fromEpochSeconds(paddedInstant.epochSeconds)
   }
 
   /**
@@ -2607,7 +2617,7 @@ abstract class AbstractFileSystemTest(
    */
   private fun Instant.maxFileSystemTime(): Instant {
     val paddedInstant = plus(200.milliseconds)
-    return Instant.fromEpochSeconds(paddedInstant.plus(2.seconds).epochSeconds)
+    return fromEpochSeconds(paddedInstant.plus(2.seconds).epochSeconds)
   }
 
   /**