+++ okio-wasifilesystem/src/wasmMain/kotlin/okio/WasiFileSystem.kt
@@ -20,10 +20,14 @@ import kotlin.wasm.unsafe.withScopedMemoryAllocator
 import okio.Path.Companion.toPath
 import okio.internal.ErrnoException
 import okio.internal.fdClose
-import okio.internal.preview1.FirstPreopenDirectoryTmp
+import okio.internal.preview1.Errno
 import okio.internal.preview1.dirnamelen
 import okio.internal.preview1.fd
+import okio.internal.preview1.fd_prestat_dir_name
+import okio.internal.preview1.fd_prestat_get
 import okio.internal.preview1.fd_readdir
+import okio.internal.preview1.fdflags
+import okio.internal.preview1.fdflags_append
 import okio.internal.preview1.filetype
 import okio.internal.preview1.filetype_directory
 import okio.internal.preview1.filetype_regular_file
@@ -31,16 +35,22 @@ import okio.internal.preview1.filetype_symbolic_link
 import okio.internal.preview1.oflag_creat
 import okio.internal.preview1.oflag_directory
 import okio.internal.preview1.oflag_excl
+import okio.internal.preview1.oflag_trunc
 import okio.internal.preview1.oflags
 import okio.internal.preview1.path_create_directory
 import okio.internal.preview1.path_filestat_get
 import okio.internal.preview1.path_open
 import okio.internal.preview1.path_readlink
+import okio.internal.preview1.path_remove_directory
 import okio.internal.preview1.path_rename
 import okio.internal.preview1.path_symlink
 import okio.internal.preview1.path_unlink_file
+import okio.internal.preview1.right_fd_filestat_get
+import okio.internal.preview1.right_fd_filestat_set_size
 import okio.internal.preview1.right_fd_read
 import okio.internal.preview1.right_fd_readdir
+import okio.internal.preview1.right_fd_seek
+import okio.internal.preview1.right_fd_sync
 import okio.internal.preview1.right_fd_write
 import okio.internal.preview1.rights
 import okio.internal.readString
@@ -52,22 +62,104 @@ import okio.internal.write
  * [WASI]: https://wasi.dev/
  */
 object WasiFileSystem : FileSystem() {
+  private val preopens: List<Preopen> = buildList {
+    // File descriptor of the first preopen in the `WASI` instance's configured `preopens` property.
+    // This is 3 by default, assuming `stdin` is 0, `stdout` is 1, and `stderr` is 2. Other preopens
+    // are assigned sequentially starting at this value.
+    val firstPreopen = 3
+
+    withScopedMemoryAllocator { allocator ->
+      val bufSize = 2048
+      val bufPointer = allocator.allocate(bufSize)
+
+      for (fd in firstPreopen..Int.MAX_VALUE) {
+        val getReturnPointer = allocator.allocate(12)
+
+        val getErrno = fd_prestat_get(fd, getReturnPointer.address.toInt())
+        if (getErrno == Errno.badf.ordinal) break // No more preopens.
+        if (getErrno != 0) throw ErrnoException(getErrno.toShort())
+
+        val size = (getReturnPointer + 4).loadInt()
+        require(size + 1 < bufSize) { "unexpected preopen size: $size" }
+        val dirNameErrno = fd_prestat_dir_name(fd, bufPointer.address.toInt(), size + 1)
+        if (dirNameErrno != 0) throw ErrnoException(dirNameErrno.toShort())
+        val dirName = bufPointer.readString(size)
+        val dirNamePath = dirName.toPath()
+        add(Preopen(dirNamePath, dirNamePath.segmentsBytes, fd))
+      }
+    }
+  }
+
+  private val relativePathPreopen: Preopen = preopens.firstOrNull()
+    ?: throw IllegalStateException("no preopens")
+
   override fun canonicalize(path: Path): Path {
-    TODO("Not yet implemented")
+    val absolutePath = when {
+      path.isAbsolute -> path
+      else -> relativePathPreopen.path.resolve(path, normalize = true)
+    }
+
+    // There's no APIs in preview1 to canonicalize a path. We give it a best effort by resolving
+    // all symlinks, but this could result in a relative path.
+    val result = resolveSymlinks(absolutePath, 0).normalized()
+
+    check(result.isAbsolute) {
+      "Canonicalize $path returned non-absolute path: $result"
+    }
+
+    return result
+  }
+
+  private fun resolveSymlinks(
+    path: Path,
+    recurseCount: Int = 0,
+  ): Path {
+    // 40 is chosen for consistency with the Linux kernel (which previously used 8).
+    if (recurseCount > 40) throw IOException("symlink cycle?")
+
+    val parent = path.parent
+    val resolvedParent = when {
+      parent != null -> resolveSymlinks(parent, recurseCount + 1)
+      else -> null
+    }
+    val pathWithResolvedParent = when {
+      resolvedParent != null -> resolvedParent.resolve(path.name)
+      else -> path
+    }
+
+    val symlinkTarget = metadata(pathWithResolvedParent).symlinkTarget
+      ?: return pathWithResolvedParent
+
+    val resolvedSymlinkTarget = when {
+      symlinkTarget.isAbsolute -> symlinkTarget
+      resolvedParent != null -> resolvedParent.resolve(symlinkTarget)
+      else -> symlinkTarget
+    }
+
+    return resolveSymlinks(resolvedSymlinkTarget, recurseCount + 1)
   }
 
   override fun metadataOrNull(path: Path): FileMetadata? {
     withScopedMemoryAllocator { allocator ->
       val returnPointer = allocator.allocate(64)
+      val preopen = preopenForPath(path) ?: return null
       val (pathAddress, pathSize) = allocator.write(path.toString())
 
       val errno = path_filestat_get(
-        fd = FirstPreopenDirectoryTmp,
+        fd = preopen.fd,
         flags = 0,
         path = pathAddress.address.toInt(),
         pathSize = pathSize,
         returnPointer = returnPointer.address.toInt(),
       )
+
+      when (errno) {
+        // 'notcapable' means our preopens don't cover this path. This will happen for paths
+        // like '/' that are an ancestor of our preopens.
+        Errno.notcapable.ordinal -> return FileMetadata(isDirectory = true)
+        Errno.noent.ordinal -> return null
+      }
+
       if (errno != 0) throw ErrnoException(errno.toShort())
 
       // Skip device, offset 0.
@@ -85,7 +177,7 @@ object WasiFileSystem : FileSystem() {
           val bufPointer = allocator.allocate(bufLen)
           val readlinkReturnPointer = allocator.allocate(4) // `size` is u32, 4 bytes.
           val readlinkErrno = path_readlink(
-            fd = FirstPreopenDirectoryTmp,
+            fd = preopen.fd,
             path = pathAddress.address.toInt(),
             pathSize = pathSize,
             buf = bufPointer.address.toInt(),
@@ -115,7 +207,7 @@ object WasiFileSystem : FileSystem() {
 
   override fun list(dir: Path): List<Path> {
     val fd = pathOpen(
-      path = dir.toString(),
+      path = dir,
       oflags = oflag_directory,
       rightsBase = right_fd_readdir,
     )
@@ -127,8 +219,15 @@ object WasiFileSystem : FileSystem() {
   }
 
   override fun listOrNull(dir: Path): List<Path>? {
-    // TODO: don't throw if the directory doesn't exist, etc.
-    return list(dir)
+    // TODO: stop using exceptions for flow control.
+    try {
+      return list(dir)
+    } catch (e: FileNotFoundException) {
+      return null
+    } catch (e: ErrnoException) {
+      if (e.errno == Errno.notdir) return null
+      throw e
+    }
   }
 
   private fun list(dir: Path, fd: fd): List<Path> {
@@ -175,22 +274,51 @@ object WasiFileSystem : FileSystem() {
         result += dir / name
       }
 
+      result.sort()
       return result
     }
   }
 
   override fun openReadOnly(file: Path): FileHandle {
-    TODO("Not yet implemented")
+    val rightsBase = right_fd_filestat_get or
+      right_fd_read or
+      right_fd_seek or
+      right_fd_sync
+    val fd = pathOpen(
+      path = file,
+      oflags = 0,
+      rightsBase = rightsBase,
+    )
+    return WasiFileHandle(fd, readWrite = false)
   }
 
   override fun openReadWrite(file: Path, mustCreate: Boolean, mustExist: Boolean): FileHandle {
-    TODO("Not yet implemented")
+    val oflags = when {
+      mustCreate && mustExist -> {
+        throw IllegalArgumentException("Cannot require mustCreate and mustExist at the same time.")
+      }
+      mustCreate -> oflag_creat or oflag_excl
+      mustExist -> 0
+      else -> oflag_creat
+    }
+    val rightsBase = right_fd_filestat_get or
+      right_fd_filestat_set_size or
+      right_fd_read or
+      right_fd_seek or
+      right_fd_sync or
+      right_fd_write
+    val fd = pathOpen(
+      path = file,
+      oflags = oflags,
+      rightsBase = rightsBase,
+    )
+    return WasiFileHandle(fd, readWrite = true)
   }
 
   override fun source(file: Path): Source {
     return FileSource(
       fd = pathOpen(
-        path = file.toString(),
+        path = file,
         oflags = 0,
         rightsBase = right_fd_read,
       ),
@@ -199,33 +327,49 @@ object WasiFileSystem : FileSystem() {
 
   override fun sink(file: Path, mustCreate: Boolean): Sink {
     val oflags = when {
-      mustCreate -> oflag_creat or oflag_excl
-      else -> oflag_creat
+      mustCreate -> oflag_creat or oflag_excl or oflag_trunc
+      else -> oflag_creat or oflag_trunc
     }
 
     return FileSink(
       fd = pathOpen(
-        path = file.toString(),
+        path = file,
         oflags = oflags,
-        rightsBase = right_fd_write,
+        rightsBase = right_fd_write or right_fd_sync,
       ),
     )
   }
 
   override fun appendingSink(file: Path, mustExist: Boolean): Sink {
-    TODO("Not yet implemented")
+    val oflags = when {
+      mustExist -> 0
+      else -> oflag_creat
+    }
+
+    return FileSink(
+      fd = pathOpen(
+        path = file,
+        oflags = oflags,
+        rightsBase = right_fd_write,
+        fdflags = fdflags_append,
+      ),
+    )
   }
 
   override fun createDirectory(dir: Path, mustCreate: Boolean) {
-    // TODO: honor mustCreate.
     withScopedMemoryAllocator { allocator ->
       val (pathAddress, pathSize) = allocator.write(dir.toString())
 
       val errno = path_create_directory(
-        fd = FirstPreopenDirectoryTmp,
+        fd = preopenForPath(dir)?.fd ?: throw FileNotFoundException("no preopen: $dir"),
         path = pathAddress.address.toInt(),
         pathSize = pathSize,
       )
+      if (errno == Errno.exist.ordinal) {
+        if (mustCreate) throw IOException("already exists: $dir")
+        return
+      }
+
       if (errno != 0) throw ErrnoException(errno.toShort())
     }
   }
@@ -236,40 +380,73 @@ object WasiFileSystem : FileSystem() {
       val (targetPathAddress, targetPathSize) = allocator.write(target.toString())
 
       val errno = path_rename(
-        fd = FirstPreopenDirectoryTmp,
+        fd = preopenForPath(source)?.fd ?: throw FileNotFoundException("no preopen: $source"),
         old_path = sourcePathAddress.address.toInt(),
         old_pathSize = sourcePathSize,
-        new_fd = FirstPreopenDirectoryTmp,
+        new_fd = preopenForPath(target)?.fd ?: throw FileNotFoundException("no preopen: $target"),
         new_path = targetPathAddress.address.toInt(),
         new_pathSize = targetPathSize,
       )
+      if (errno == Errno.noent.ordinal) {
+        throw FileNotFoundException("no such file: $source")
+      }
+
       if (errno != 0) throw ErrnoException(errno.toShort())
     }
   }
 
   override fun delete(path: Path, mustExist: Boolean) {
-    // TODO: honor mustExist.
     withScopedMemoryAllocator { allocator ->
       val (pathAddress, pathSize) = allocator.write(path.toString())
+      val preopenFd = preopenForPath(path) ?: throw FileNotFoundException("no preopen: $path")
 
-      val errno = path_unlink_file(
-        fd = FirstPreopenDirectoryTmp,
+      var errno = path_unlink_file(
+        fd = preopenFd.fd,
         path = pathAddress.address.toInt(),
         pathSize = pathSize,
       )
+      // If unlink failed, try remove_directory.
+      when (errno) {
+        Errno.noent.ordinal -> {
+          if (mustExist) throw FileNotFoundException("no such file: $path")
+          return // Nothing to delete.
+        }
+
+        Errno.perm.ordinal,
+        Errno.isdir.ordinal,
+        -> {
+          errno = path_remove_directory(
+            fd = preopenFd.fd,
+            path = pathAddress.address.toInt(),
+            pathSize = pathSize,
+          )
+        }
+      }
       if (errno != 0) throw ErrnoException(errno.toShort())
     }
   }
 
   override fun createSymlink(source: Path, target: Path) {
     withScopedMemoryAllocator { allocator ->
+      val sourcePreopen = preopenForPath(source)
+        ?: throw FileNotFoundException("no preopen: $source")
+
+      // Always create symlinks relative to their source. Absolute symlinks are trouble because the
+      // absolute paths used by WASI are different from the absolute paths on the host file system.
+      val sourceParent = source.parent
+        ?: throw IOException("unexpected symlink source: $source")
+      val targetRelative = when {
+        target.isRelative -> target
+        else -> target.relativeTo(sourceParent)
+      }
+
       val (sourcePathAddress, sourcePathSize) = allocator.write(source.toString())
-      val (targetPathAddress, targetPathSize) = allocator.write(target.toString())
+      val (targetPathAddress, targetPathSize) = allocator.write(targetRelative.toString())
 
       val errno = path_symlink(
         old_path = targetPathAddress.address.toInt(),
         old_pathSize = targetPathSize,
-        fd = FirstPreopenDirectoryTmp,
+        fd = sourcePreopen.fd,
         new_path = sourcePathAddress.address.toInt(),
         new_pathSize = sourcePathSize,
       )
@@ -278,29 +455,58 @@ object WasiFileSystem : FileSystem() {
   }
 
   private fun pathOpen(
-    path: String,
+    path: Path,
     oflags: oflags,
     rightsBase: rights,
+    fdflags: fdflags = 0,
   ): fd {
     withScopedMemoryAllocator { allocator ->
-      val (pathAddress, pathSize) = allocator.write(path)
+      val preopenFd = preopenForPath(path) ?: throw FileNotFoundException("no preopen: $path")
+      val (pathAddress, pathSize) = allocator.write(path.toString())
 
       val returnPointer: Pointer = allocator.allocate(4) // fd is u32.
       val errno = path_open(
-        fd = FirstPreopenDirectoryTmp,
+        fd = preopenFd.fd,
         dirflags = 0,
         path = pathAddress.address.toInt(),
         pathSize = pathSize,
         oflags = oflags,
         fs_rights_base = rightsBase,
         fs_rights_inheriting = 0,
-        fdflags = 0,
+        fdflags = fdflags,
         returnPointer = returnPointer.address.toInt(),
       )
+      if (errno == Errno.noent.ordinal) {
+        throw FileNotFoundException("no such file: $path")
+      }
       if (errno != 0) throw ErrnoException(errno.toShort())
       return returnPointer.loadInt()
     }
   }
 
+  /**
+   * Returns the preopen whose path is either an ancestor of [path], or whose path [path] is an
+   * ancestor of.
+   *
+   * If [path] is an ancestor of our preopen, then operating on the path will ultimately fail with a
+   * `notcapable` errno.
+   */
+  private fun preopenForPath(path: Path): Preopen? {
+    if (path.isRelative) return relativePathPreopen
+
+    val pathSegmentsBytes = path.segmentsBytes
+
+    return preopens.firstOrNull { preopen ->
+      val commonSize = minOf(pathSegmentsBytes.size, preopen.segmentsBytes.size)
+      preopen.segmentsBytes.subList(0, commonSize) == pathSegmentsBytes.subList(0, commonSize)
+    }
+  }
+
   override fun toString() = "okio.WasiFileSystem"
+
+  private class Preopen(
+    val path: Path,
+    val segmentsBytes: List<ByteString>,
+    val fd: fd,
+  )
 }